requirements.txt
# requirements.txt
requests==2.26.0
pandas==1.3.3
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
python-dotenv==0.19.0
logging==0.5.1.2